Added include of serializer.h to pfor.h
Simple workaround for a problem in AllScale with the latest API version.

Including serializer.h directly in pfor.h (even though we do not use it
here in any way) fixes a problem introduced in a previous change
(8670eb5), which caused it to be included transitively via prec.h,
treeture.h, reference/treeture.h, profiling.h, and task_id.h.

Including it transitively via the intercepted prec.h causes problems, as
classes declared in serializer.h then get prec.h annotated as include
location. This is by design and how it should work for system headers.
However, in this case we would prefer it to get the correct header
attached, but for this we would need to implement a new mechanism in the
header lookup mechanism.

The problem is, that the AllScale compiler checks that nothing has prec.h
(and some others) attached. Since serializer.h is now included here directly
it gets the correct header attached and the assertion does not trigger.